app/src/main/java/com/felina/fianreqres/ui/MainViewModel.kt
package com.felina.fianreqres.ui

import android.content.Context
import androidx.lifecycle.LiveData
import androidx.lifecycle.ViewModel
import androidx.lifecycle.ViewModelProvider
import androidx.lifecycle.viewModelScope
import androidx.paging.PagingData
import androidx.paging.cachedIn
import com.felina.fianreqres.data.UserRepository
import com.felina.fianreqres.di.Injection
import com.felina.fianreqres.network.DataItem

class MainViewModel(userRepository: UserRepository) : ViewModel() {

    val user: LiveData<PagingData<DataItem>> =
        userRepository.getUsers().cachedIn(viewModelScope)
}

class ViewModelFactory(private val context: Context) : ViewModelProvider.Factory {
    override fun <T : ViewModel> create(modelClass: Class<T>): T {
        if (modelClass.isAssignableFrom(MainViewModel::class.java)) {
            @Suppress("UNCHECKED_CAST")
            return MainViewModel(Injection.provideRepository(context)) as T
        }
        throw IllegalArgumentException("Unknown ViewModel class")
    }
}